@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 1.3.0+11
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   cached_network_image: ^3.0.0
