@@ -1,3 +1,5 @@
+// @dart=2.9
+
 import 'package:finale/services/lastfm/common.dart';
 import 'package:flutter/material.dart';
 
@@ -9,7 +11,7 @@ class WikiComponent extends StatelessWidget {
   Widget _buildPage(BuildContext context) => Scaffold(
       appBar: AppBar(centerTitle: true, title: Text('Wiki')),
       body: ListView(padding: EdgeInsets.all(10), children: [
-        Text(wiki.content.substring(0, wiki.content.indexOf('<a')).trim()),
+        Text(wiki.content),
         SizedBox(height: 10),
         SafeArea(
             child: Text('Published ${wiki.published}',
@@ -18,8 +20,7 @@ class WikiComponent extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => ListTile(
-      title: Text(wiki.summary.trim(),
-          maxLines: 3, overflow: TextOverflow.ellipsis),
+      title: Text(wiki.summary, maxLines: 3, overflow: TextOverflow.ellipsis),
       trailing: Icon(Icons.chevron_right),
       onTap: () {
         Navigator.push(context, MaterialPageRoute(builder: _buildPage));
